Implementa gerenciamento de edital

+++ src/main/java/portal/editais/repository/RegiaoImediataRepository.java
@@ -1,12 +1,12 @@
 package portal.editais.repository;
 
-import java.util.Optional;
 import java.util.List;
+import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import portal.editais.entity.RegiaoImediata;
 
-public interface RegiaoImediataRepository extends JpaRepository<RegiaoImediata, Long> {
-    Optional<RegiaoImediata> findByIdAndAtivoTrue(Long id);
+public interface RegiaoImediataRepository extends JpaRepository<RegiaoImediata, Integer> {
+    Optional<RegiaoImediata> findByIdAndAtivoTrue(Integer id);
 
     boolean existsByCodigo(String codigo);
